Make password autogeneration option obvious
Re-added the password auto-generation checkbox back to the password
input group (now just called "Auto"). The checkbox is always visible.

To counteract the used space, we now selectively hide the signal (for
SSID) and padlock (for password) icons on mobile (xs breakpoint) so as
to create more space for those text fields.

Also, did a bit more cleanup.